backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pydantic>=2.0
httpx>=0.27
PyPortfolioOpt>=1.5
yfinance>=0.2.37
pandas>=2.0
numpy>=1.26
python-dotenv>=1.0
anthropic>=0.25.0
